[Partial fix for #5] Added early version of a real track view.
An early version of the track view has been created. While this is in fact already using seperate images for the sections and does take its track data from the in-memory data structure, how it is using the data structure has to be overhauled majorly in the future. Also, the track is currently defined inside the track view for testing purposes.

It's already possible to pan the view by holding the right mouse button and the display is also double-buffered.