[Partial fix for #5] Implemented first working version of the dynamic track view.
This is the first working version of a track view that creates a track dynamically from a set of tiles.

So far, it has only been tested with 90-degree turns and there's still an abundance of unused code, hacky workarounds and undocumented sections. This will change soon.

A new bean has been introduced, the TilesetRepository (name probably going to change). It's used to store the set of available tiles of various tilesets. The class TrackSection might go away as well.