[Related to ticket #5] Added a setter for the track name to the Track class and improved documentation.
A setter has been added for the track name of the Track class (so it can be changed while running in designer mode) and the used tile set is now stored and can be retrieved.

The javadoc documentation has been improved.

NOTE: The class is no longer Serializable, as the loading and saving is done manually via member methods anyway.