[Partial fix for ticket #5, related to ticket #6] The current Track of the !TrackWidget can now be changed and the !TrackWidget is notified about modifications of the track.
The !TrackWidget now expects a Track object in the constructor that will be used to initialize the Widget with. The current track can also be changed using the setTrack() method.

Additionally, the !TrackWidget now automatically registers itself as a listener on the track (both when the track is assigned via the constructor as well as when it is changed using setTrack()) and will update itself when the track changes.