[Partial fix for ticket #5] Replaced Graphics2DRenderer with the 'Java2D for SWT' library.
Added the [http://www.holongate.org/holongate-2.0/ Java2D for SWT] library to the repository and included it in the build path.

Adjusted the !TrackWidget to use the library for rendering the track. This improves performance greatly if a native acceleration library is available. (Windows and Linux (GTK/Motif) libraries are included in the repository now.)

The [http://www-128.ibm.com/developerworks/java/library/j-2dswt/ Graphics2DRenderer] is no longer required and has been removed.

NOTE: "-Djava.library.path=${workspace_loc:JACAVI}/libs" has to be added to the "VM arguments" of the launch configuration to make use of the native acceleration libraries.