[Partial fix for ticket #6, related to ticket #19] The track designer now retrieves the available tiles from the tileset repository.
The track designer now retrieves the available tiles from the tileset repository. It's currently using only the debug tileset, because the class doesn't know which tileset the user wants yet.

The preview images are loaded as SWT images, separately from the Java2D images stored in the tileset repository. It could be discussed whether it would make more sense to convert the Java2D images to SWT images (which technically simple), as the Java2D are already in memory anyway.

Some modifications were necessary to make this possible:
 * Tiles now have a display name.
 * The Tile now provides getters for the display name and the filename.
 * The !TilesetRepository now stores the tiles of a specific tileset in a SortedMap instead of a Map, so the ordering used in the tiles.xml will also be used in the track designer.

 NOTE: The application will crash if no display name of a tile (attribute name on the tiles element) is defined in the tiles.xml.