[Partial fix for ticket #5] Proof-of-concept for track rendering engine reimplementation.
The rendering engine has been almost completely rewritten, making heavy use of the Java2D affine transforms. No custom calculation is necessary anymore. This is an intermediate commit that is not yet cleaned up or refactored.

Among other things, this made it possible to do hit detection. A track section is now selected when it is double-clicked.