[Partial fix for ticket #5] Refactored and cleaned up the !TrackWidget.
The !TrackWidget has been cleaned up, refactored from ground up and javadoc'ed.

The selection algorithm has been fixed and is more intuitive now. Selections are attempted to be preserved, modified if possible and only removed if necessary.

Selection is now done using a single click of the left mouse button and the selection can be removed by clicking anywhere outside the tiles.

The !TrackDesigner had to be modified because the !TrackWidget constructor no longer throws an exception (previously couldn't either, but the throws clause was still there).