[Partial fix for ticket #5, related to ticket #6] Improved code for removing sections in the designer and prevented the initial tile from being removed.
The code for removing tiles in the !TrackDesigner has been moved into the Track's removeSection() method.

The initial tile is now properly determined from the tiles.xml and used to initialize a new thread. It can no longer be removed (removeSection() throws an exception) and it's no longer part of the list returned by !TilesetRepository.getAvailableTiles().

Additionally, a bug has been fixed that prevented the !TrackWidget from rerendering after a section insertion if no selection was active.