[Fix and cleanup for ticket #19] Restored the debug tiles and updated the tile definition file to the latest format.
Restored the debug tiles that were accidentally removed in [135]. Also added them back to the tiles.xml.

Initial tiles are now mandatory, so an initial tile was added to the digital tileset (currently using the same image as the normal straight). The demo track was extended by a reference to this initial tile.