[Partial fix for ticket #25, related to ticket #19] Added support for specifying the car slots in the tile definition and added debug code for displaying the slots and moving objects around on them.
The tiles.xml now requires two slots to be specified for each tile. A slot consists of one or more parts of various types. The slot parts are used in the order specified.

So far, two different slot part types have been implemented. The element <line> can be used to define a slot part that goes straight from a point A to a point B. The element <quad-bezier> can be used to define a [http://en.wikipedia.org/wiki/Bezier_curves quadratic Bézier curve].

The length attribute defines how many 'steps' are used to traverse a slot part and need to be modeled very accurately after the real track. It is also used to assign each position on the track a unique scalar value, thereby making locations very simple to store and modify.

Purely for demonstration purposes, the track widget currently draws the slots (in blue and yellow) and moves carets along them (in red). Determining the direction of an object on a slot has not yet been implemented.

NOTE 1: The application currently crashes when it is terminated while the track widget is still open. This is probably because the widget's dispose handler isn't invoked before the widget is disposed. I haven't found a way to fix that yet.

NOTE 2: The CPU usage is still rather high, I'm sure optimisations are possible.