[Partial fix for ticket #25] Fixed a bug that occasionally caused the application to crash during exit and added a simple frame counter.
The application occasionally crashed during exit. This was caused by an anonymous class (Runnable) that was asynchronously pushed onto the SWT event thread by the timer. This Runnable was sometimes executed after the widget had already disposed. It is now confirming that the widget is not yet disposed before executing its code (triggering a repaint).

Additionally, a simple frame counter was added. It normally shouldn't go above the gametick rate (currently 20/s), but it can be used to compare performance when higher gametick rates are used.