[Partial fix for ticket #6] Exception Handling and scaled Images in the Toolbar
The Exceptions are handled by a simple ErrorDialog.

The Tile images are now scaled. Now there are a few Toolbars embedded in one CoolBarItem, so they could be scaled. After experimenting a lot today, I think it looks a bit better than the last version, but it's not the finish one.