[Final fix for ticket #5] Implemented an inner control system for the track view widget and added panning, rotation and zooming controls.
When hovering over the track view widget, it now shows a set of controls for controlling the viewport of the track view (scrollers for panning, buttons for rotation and zooming).

Icons have been included from the famfamfam silk set.

A new util class, !GrayscalableImage, has been added for images that need to be shown coloured and in greyscale. This is necessary because converting an image to greyscale takes some time and caching this improves performance a lot.