[Partial fix for ticket #6 and #42] Track Designer redesigned
The Trackdesigner now instantiates only the Track Widget. The Tiles to enlarge the track are outsourced in the Tile Explorer View (maybe the name has to be reconsider). This view is always updated if the Active Editor change.
The Tiles aren't more embedded into a Coolbar, but now in several Buttons, which could now be scrolled.