[Partial fix for ticket #39] Created an !InputDeviceManager (bean) and added a preliminary integration with the player settings dialog.
The !XyzDeviceController classes have been renamed to !XyzDevice because they now represent instances of a device. (If the device is used for more than one player, there are also multiple instances of the class.)
They are all subclasses of the abstract class !DeviceController (formerly an interface) which contains common code, like managing the ID and the name. The ID will be necessary to track modifications of input devices that are already assigned to a player. That makes the Devices enum superfluous.

The !PlayerSettingsDialog does use the !InputDeviceManager to fill its combo boxes but has not yet been completely adjusted.