[Fix for ticket #40] Implemented support for saving tracks.
Tracks can now be saved to a file by invoking the Track's saveToXml() method. This also has been integrated into the track designer.

Additionally, a bug has been fixed that caused the track name to be read incorrectly (trackName was set to the tileset name, instead of the actual name specified in the file).