[Fix for ticket #55] Implemented thrust gauges to be shown on the !TrackWidget.
The upper left corner of the !TrackWidget now displays thrust gauges. These are currently filled with data directly from all available '''input devices''' and not from the '''players''' as it should be. Also, the devices are polled at the time of drawing instead of the polling being triggered by the !RaceEngine.

To make this temporarily work, the !InputDeviceManager now calls the !DeviceController's initialize() method when a new device is added. This is just a hack and has to be removed because it can cause devices to be initialized mulitple times and initializes devices long before they will be used in the final version (the initialization should probably take place when the race is started).
Additionally, the KeyboardDevice now returns a fixed signal when polled (always 50% thrust) for testing.

Besides, the !TrackWidget now uses a customized font for displaying player (currently: device) names and the frame counter.