[Fix for ticket #56, partial fix for ticket #53] Fixed !NullPointerException during scrolling.
I'm not entirely sure which race condition caused this, but in certain situations a !NullPointerException was thrown when moving the cursor away from an inner !TrackWidget control while holding it. It was somehow caused by the timer thread for the !TimerTask !ClickEventRepetitionHandler communicating with the event dispatcher thread over the field hoveredInnerControl. This problem was worked around by passing the current hoveredInnerControl into the !ClickEventRepetitionHandler during construction. The !ClickEventRepetitionHandler only works on its own state now and in the worst case invokes the requested change once more while the mouse button has already been released.

During investigation, another bug was found that caused the inner control clicks to be also checked for track section selections. That has been fixed, clicks on inner controls now have priority and the track selection change is only evaluated if no inner control click could be detected.

Additionally, the scrollers' click detection shape is now a little larger, making it easier to hit them. The previous shape was more exact (having the same shape as the scroller), but also did not take the brush width into account.