[Related to ticket #60] fitted the RaceEngine to our purpose and added an RaceView interface
Now the RaceEngine is open to get accessed from outside and holds an instance of RaceView given on startRaceTimer. In raceTimerInterval configured in jacavi.properties and over configurationBeans.xml the RaceView will be repainted.

I created the Interface just for implementing the RaceEngine don't know if there is further need if the real RaceView exists