[Fix for ticket #62] Added support for two different modes of operation in the !TrackWidget.
The differences are documented in the javadoc of the !TrackWidgetMode enum.

The constructor invocations in the !TrackDesigner and the !TrackView have been modified to use MODE_DESIGN and MODE_RACE, respectively.

Note that the animation thread has to be removed as soon as the animation is triggered by the !RaceEngine!