[Partial fix for ticket #39] Major rewrite of the input device settings dialog.
The device types are now configured in a tab folder. The 'four sections' design didn't leave enough place for proper configuration. Additionally, a list of all configured devices has been added at the bottom to give a quick overview which devices can be used.
Most of this functionality has been factored out into an abstract class (!AbstractSettingsDialog) so it can be reused for the connector configuration dialog.

The !InputDeviceManager has been overhauled and now actually manages the devices instead of just being a more comfortable collection class. That also allowed to do proper one-time checks on whether the external libraries (for game controllers and Wiimotes) are available. If they are not, the corresponding sections are now completely deactivated and just refer the user to the logs.

Some other modifications were necessary to get all this working:
 * the !GameControllerDeviceManager no longer catches the !UnsatisfiedLinkError itself, as the !InputDeviceManager must know about this
 * the playersBean is now initialized empty
 * the !CarController has been made an abstract class and does the ID and name management (this functionality has been pulled up from the !DeviceController, as it will be needed for driving agents anyway)
 * the !MouseDevice returns a debug signal as the !KeyboardDevice already did