[Related to ticket #2] Merged, consolidated and cleaned up launch configurations.
There are now only two launch configurations again, both with a clear purpose:
 * JaCaVi (accelerated) defines the VM argument java.library.path and the environment variables PATH and LD_LIBRARY_PATH (all on WORKSPACE/libs) and will hopefully work on both Windows and Linux
 * JaCaVi (no external libs) defines none of these arguments and simulates a platform were none of the native libs are available (MacOS, for example) -- it can be used to make sure JaCaVi still runs on these platforms and handles the missing libraries gracefully

In both configurations the plug-in list has been recreated from scratch by deselecting all, then selecting the JaCaVi plug-in and letting Eclipse figure out all plugins that are missing (optional dependencies excluded). In the future, care should be taken to increase this set with care and only if absolutely necessary. (Particularly because the launch configurations can hardly be diffed -- Eclipse seems to reorder the list each time.)