[Fix for ticket #84, related to #39 and #63] Fixed exception that was thrown when the input settings dialog was closed with the X-button in the top right.
This was caused by the cleanup code being in the okPressed() handler, which was not called when the X-button was clicked.

The X-button has now been replaced with a 'Close' button as suggested in the ticket to indicate that all changes are immediate and there's no difference whether the dialog is closed with the X or with the Close button. The cleanup code has been moved to the close() method.

Also, the preview update timer now checks if its widget (the progress bar) is still available before modifying it.

Additionally, the layout of the 'mouse' tab has been fixed.