[Partial fix for ticket #13] Added an initial KeyboardDevice Conroller
The KeyboardDevice works with a global listener which is added to the the Display. For a clean solution the listener has to be added when the race starts and removed, when the race end up.(cleanup method)

To make it configurable I also added a KeyboardLayout class, where the keys could be bind to an action. This layout may be added to the KeyboardDevice constructor.
(The current version has a static KeyboardLayout.Default member)