[Partial fix for ticket #12] Added an initial MouseDevice Conroller
The MouseDevice works with a global listener like the KeyBoaed Device(see #13). 
For a clean solution the listener has to be added when the race starts and removed, when the race end up.(cleanup method)

Maybe we really have to integrate an additionally MouseView, where the mouse has the focus on. Actually the current version works fine, but the MouseEvents of the TrackWidget compete against.