[Partial fix for tickets #12 and #13] A preview for mouse and keyboard layouts has been added to the !InputDeviceSettingsDialog.
The implementation is rather hackish and probably still contains some bugs. Cleanup and refactoring wouldn't hurt.

Additionally, a bug in the mouse device has been fixed that caused it to 'stick' when unhookListener() was invoked while the mouse button was pressed.

Also, the device states of both mouse and keyboard are now reset on hookListener() and some warnings have been fixed.