[Fix for ticket #82, related to ticket #2] Enhanced position information stored about each car.
The position of the car is no longer a simple int, it is now stored in a CarPosition instance.

What was formerly called a !TrackPosition has now been renamed to !CarScreenPosition for clarification. It is simply a helper class for passing some drawing information around and shouldn't be used outside the TrackWidget.

In the context of this change, it was diagnosed that the race bean had never been properly set (e.g. the players were sometimes accessed directly via the players bean and sometimes via the race bean) which caused a rather huge related set of changes:
 * The race bean has been removed.
 * Classes that used the race bean so far are now directly using the players bean or the !RaceEngine bean.
 * The validators get the players list as a parameter.
 * The !RaceEngine methods startRaceTimer() and stopRaceTimer() have been renamed to startRace() and stopRace() because they have long been doing much more than just starting the timer.
 * The active track is no longer taken for the active editor's !TrackWidget but directly from the editor.
 * The !DebugTDA now supports moving the cars more than one lap.