[Partial fix for ticket #22] Renamed and moved the hookListener() and unhookListener() methods of the DeviceController up in hierarchy.
The hookListener() and unhookListener() methods of the DeviceController have been moved up in hierarchy (to the CarController), so they can be called on driving agents as well. That fixes the exception that was thrown during race start and termination if a driving agent was involved.

They have also been renamed to preRace() and postRace() to be able to use them for more general tasks in the future.