[Related for ticket #1] Various renamings and cleanups in the !CarController interface.
preRace() and postRace() have been renamed again to activate() and deactivate(). deactivate() is required to be allowed to call multiple times without negative effects.

The initialize() method was removed as it is no longer used.

The cleanup() method is now optional (as are activate() and deactivate()) to override and by default just calls deactivate().