[Related to ticket #92] implemented front and back light on/off support for Bluerider
I also maped wiimote button 1 on frontlight and wiimote button 2 on backlight.
I figgured out that signals like light on/off are a bit more special because th shouldnot get lost in timer like speed signals. I think we have to make shure that a ControllerDevice should only work with one instance of ControllerSignal. So we can get shure that such signals will arrive the RaceEngine and we could then set them back in the RaceEngine.