[Fix for ticket #88] Implemented a proof-of-concept manager for the cars and modified !TrackWidget code to draw the actual cars.
Cars are now defined in /cars/cars.xml. They can be accessed via the carRepositoryBean. The !TrackWidget has been modified slightly to draw them on the track.

Additionally, a bug in the DebugTDA has been fixed that raised NullPointerExceptions when the FeedbackSignal did not contain a g-force component and the gametick interval has been lowered.