[Partial fix for ticket #50] Added proof-of-concept implementation of persistent editors. (Needs review.)
In the !ApplicationWorkbenchAdvisor, new string preferences 'editor0' to 'editorN' are now stored with the filenames as values. The following one is removed from the preferences store to prevent loading old entries.

During startup, these preferences are read and the corresponding editors are restored (until an empty entry is found).

The !TrackDesignerInput now stores the filename, but the implementation seems to be buggy -- particularly when newly created tracks are involved.

Also removed the demo.track.xml added in the last commit which seems to have been commit accidentally and doesn't even load.