[Partial fix for tickets #49 and #82] Intermediate commit for the implementation of lane change support.
The !CarPosition no longer stores a global int to specify the car position on the track. Instead, it now stores a !TrackSection and an int for the position of the car on the given section.

The tiles.xml now supports two ways of storing the !LaneSections of a Lane. The new method is to specify three groups of !LaneSections (and specifying the attribute isLaneChange="true" in the <lane> element):
 * sections inside <common/> are always taken (first)
 * sections inside <regular/> are taken after the common ones if no lane change was triggered
 * sections inside <change/> are taken after the common ones if a lane change was triggered
If the isLaneChange attribute is not set, lane sections are accepted right inside the <lane> element as it has been up to now.

The Lane's getLength() method and the Track's getLaneLength() methods have been removed as the length now depends on whether the lane change is triggered. The Lane's getStepPoint() method and the Track's determineScreenPositionFromPosition() have been slightly modified (incomplete).

The method moveSteps() on the !CarPosition class has been adapted to the new way of storing lanes, but it doesn't support lane changes yet.

Two example tiles have been added to the debug tileset.

ATTENTION: The lane change support is incomplete, tiles with lane changes are not handled correctly right now and lane changes is *not* possible!