[Related to ticket #1] Fixed problem with property placeholders in the application context.
The !ContextLoader now uses !ApplicationContext instead of !XmlBeanFactory which is recommended by the Spring team and fixes the main issue.

However, some modifications were necessary to make this work:
 * Bean class constructors may (no longer?) use getBean(), so the !ConnectorConfigurationManager now gets its !ConnectorFactory instance injected.
 * The Lib42FeedbackManager is now lazy-init to avoid it being loaded on platforms where Lib42 isn't available.
 * The !TilesetRepository and !CarRepository are now lazy-init to avoid them being loaded when the test cases are run -- which causes an unknown problem while loading the tilesets.

The configurationBeans file has been removed and the declaration of the propertyConfigurer has been moved into the applicationContext.

The gametick interval is now taken from the jacavi.properties file and the path to the tiles and cars definition has been moved there as well.

NOTE: Two test cases are still failing, but they apparently have been for some time.