[Partial fix for tickets #49 and #82] Implemented support for setting the exit lane of a lane, the lane a car will be on after leaving a tile. [MISSING PART]
The exit lane can be specified by giving the lane index (0-based) on the exitLane attribute of the <lane> element. This will be considered the regular exit lane. In the case of lane changing tiles, the same attribute can also be specified on the <regular> and <change> elements. The attribute on <regular> will overwrite the attribute on the <lane> element. The attribute on <change> will define the exit lane of the lane change path.
The attributes are entirely optional and the default is always to stay on the same lane.

The demo_crossing track has been enhanced by a crossover tile demostrating this feature.

NOTE: Added image file missing in the last commit.