[Partial fix for ticket #22] Intermediate commit on the way to working driving agent support.
The !InputDeviceManager has been renamed to !CarControllerManager (and moved in hierarchy) because it also stores the driving agent instances now.

The !PlayerSettingsDialog has been adapted to support driving agents. Its UI is broken now, but it needs to be revised anyway.

The !TrackWidget is now showing the thrust gauges for all players again (including driving agents).