[Fix for ticket #98] Cars are now drawn over all track tiles.
This removes the drawing artifacts happening at the tile edges. However, it also removes the 'feature' of cars being drawn below tile 'bridges'. This might have to be investigated again, but for now, it's probably an improvement.

Also removed old debug car and fixed a bug in the 'cloverleaf' tile of the debug tileset.