[Partial fix for ticket #125, related to ticket #53] Redraws are no longer triggered if the !TrackWidget isn't already idle.
Before pushing a new redraw event to the event queue, the !TrackWidget now checks whether it's already working on a frame. That should reduce the problem of the refresh events for the !LapView not getting through. (In the latest tests, it still took up to 2s, unfortunately.)

The dropped frames are shown on the frame counter.