[Partial fix for ticket #102] Implemented showing off-track cars semi-transparent.
Cars that have left the track are now shown semi-transparent (currently 30% opaque) at the position they left the track. This was implemented by adding another transformation to the !MultiStyleImage, which also caches the transformation results.

The cars had to be converted because the !MultiStyleImage expects 24- or 32-bit images, and they were 8-bit palette-based. While doing so, the old images have been moved into a subfolder and the main transparency of all images has been improved by hand.