[Partial fix for ticket #14] Added some workarounds to avoid the traffic lights crashing the driving agents.
The driving agents can't be poll()ed until they are activated(). However, with the traffic lights, the !TrackWidget is drawn (and therefore poll()s the controller) before the race has been started (and therefore the controllers activate()d).

Some IFs have been added to avoid in crash in that situation.