[Partial fix for ticket #156] Refactored driving agent architecture.
A !DrivingAgentController supertype for !ScriptController has been created, which will also be used for the !ExternalController. The !ScriptController no longer contains the actually Jython/Groovy support, that has been moved into separate classes now.

In the process, !CarControllerManager.removeDevicesByType() and .getInputDevicesByType() finally check the whole type hierarchy and not just the direct superclass.