[Fix for ticket #152,#157,#149] fixed starting point problem on reset, generalized functionality on CarController.reset().
Now reset functionality etc. is back in the RaceEngine so we have a better workflow and TDA's have a bit less functionality. This was reached to give the TDA an reset function to reset saved speed and acceleration if car crashes. I think its good and fine if all race workflow logic is generalized in the RaceEngine so the TDA only has to approximate position by physics etc. Maybe we can outsource functionality currently in TrackDataApproximator like calculation of steps between checkpoints later some where in a TrackDataInformationManager initialized by RaceEngine before raceStart so TDA can get its information about track etc concentrated from TrackDataInformationManager because currently intelligence informations about the track are not easy to get. 

We need a reset on CarController because the deactivate function disables the Controller listeners for example keyboard lsiteners so a reset push button could not be detected further the RaceEngine couldnt trigger the ControlerSignal.isReset()and the car is out of race for ever. Added  javadoc to CarController reset().