[Related to ticket #87] DLL's and SO's now loaded by OSGi
We solved the problem of dependent libraries by calling System.loadLibrary(), when the application starts. We assume, that the ClassLoader register the dll from this time on and so other dll's which depends on it have no problems anymore.

This has the effect, that we do not need to set the java.library.path and PATH in our Run Configurations anymore.

Further informations about OSGi and Native Librarys at http://www.rcpapps.org/articles/Eclipse2VM.html#Native_code_and_classloaders