[Partial fix for ticket #154] The driving agents now receive information about them, their competitors and the track.
This was carried out by adding two parameters to the activate() method of the !CarController, the track and the list of players.

Additionally, a getLastSignal() method was introduced on the !CarController. This is used by the !TrackWidget to draw the thrust gauges and may actually be called before the first poll() (due to the way the race start initialization phase is performed) but never before activate(). The default implementation just invokes poll().

Using the now available data, a new driving agent has been introduced that cares a little more about the track he's driving on.