[Related to ticket #47] added an JaCaVi actors Use Case
I also added an folder png under vpproject for holding png images of the diagrams there. The Problem of exporting images from vp diagrams is with the community version, that there are watermarks in it. So i made / make screenshots of the diagrams and add them to the png folder.

Also added some diagrams in png